==> notebooks/01_API's.ipynb <==
import pandas as pd
import sys
sys.path.append("../")
from src import soporte_api as saw
import json

# API WALLAPOP

producto = "Play Station 5"
precio_min = "200"
precio_max = "350"

datos_wallpop = saw.busqueda_wallapop(producto, precio_min, precio_max)

df_datos_wallapop = pd.read_json("../data/resultado.json")
df_datos_wallapop_limpios = saw.procesar_datos_wallapop(df_datos_wallapop)

saw.filtrar_playstation5("../data/datos_wallpop_limpios_clean.csv", "../data/datos_wallapop_filtrados_definitivos.csv")

# API VINTED

datos_vinted = saw.buscar_vinted("Play Station 5")

df_datos_vinted = pd.read_json("../data/vinted_ps5.json")
df_datos_vinted_limpios = saw.procesar_datos_vinted(df_datos_vinted)
