==> notebooks/02_Scrapeo.ipynb <==
import sys
sys.path.append("../")
from src import soporte_scrap as ssc
import pandas as pd
from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from bs4 import BeautifulSoup
import time
import numpy as np
import pandas as pd

# Scrapero Cash Converters

url = "https://www.cashconverters.es/es/es/comprar/videojuegos-y-consolas/consolas/consola-ps5/?cgid=5292&q=Play+Station+5&start=0&sz=72"

datos_cash_conerters = ssc.traer_productos_cash_converter(url)

df_datos_cashconverters = pd.DataFrame(datos_cash_conerters)
df_datos_cashconverters.to_csv("../data/datos_cash_converters.csv")

# Scrapeo mercado libre

url = "https://listado.mercadolibre.com.mx/consolas-videojuegos/consolas/sony/playstation-5/usado/playstation-5_NoIndex_True#applied_filter_id%3DBRAND%26applied_filter_name%3DMarca%26applied_filter_order%3D2%26applied_value_id%3D995%26applied_value_name%3DSony%26applied_value_order%3D1%26applied_value_results%3D209%26is_custom%3Dfalse"
df_mercado_libre = ssc.traer_productos_mercado_libre(url)

columnas_a_eliminar = ["descripcion", "precio_mxn", "precio"]
df_mercado_libre = df_mercado_libre.drop(columnas_a_eliminar, axis=1, errors='ignore')

df_mercado_libre.to_csv("../data/datos_mercado_libre.csv")
