==> notebooks/03_BBDD.ipynb <==
import psycopg2
import pandas as pd
import sys
sys.path.append("../")
from src import soporte_bbdd as sbbdd

## Conexión

conn = psycopg2.connect(
    host="localhost",
    database="proyecto_5",
    user="postgres",
    password="admin"
)
cursor = conn.cursor()

## Creación tablas

sbbdd.create_tables(cursor, conn)

## Insercción datos

sbbdd.bulk_insert_from_csv('../data/datos_mercado_libre.csv', 'datos_mercado_libre', conn)


sbbdd.bulk_insert_from_csv('../data/datos_cash_converters.csv', 'datos_cash_converters', conn)

sbbdd.bulk_insert_from_csv('../data/datos_vinted_filtrados.csv', 'datos_vinted_filtrados', conn)

sbbdd.bulk_insert_from_csv('../data/datos_wallapop_filtrados_definitivos.csv', 'datos_wallpop_limpios', conn)

conn.rollback()

# Cerrar la conexión
cursor.close()
conn.close()
