==> notebooks/04_Filtros_y_notificaciones.ipynb <==
import psycopg2
import pandas as pd
import sys
sys.path.append("../")
from src import soporte_wallapop as swallapop
from src import soporte_vinted as svinted
from src import soporte_cash_converters as scash
from src import soporte_mercado_libre as smerc

He creado unas tablas vacías con la misma estructura que las tablas para introducir las que nos interese:

CREATE TABLE wallapop_interesante AS
SELECT * FROM datos_wallpop_limpios WHERE 1=0;


CREATE TABLE vinted_interesante AS
SELECT * FROM datos_vinted_filtrados WHERE 1=0;



CREATE TABLE cash_converters_interesante AS
SELECT * FROM datos_cash_converters WHERE 1=0;



CREATE TABLE mercado_libre_interesante AS
SELECT * FROM datos_mercado_libre WHERE 1=0;

Decido hacer la media de venta de mi producto (ps5), para saber que productos me pueden interesar comprar

conn = psycopg2.connect(
    host="localhost",
    database="proyecto_5",
    user="postgres",
    password="admin"
)
cursor = conn.cursor()

conn.rollback()

## Wallapop


cursor = conn.cursor()

media_productos_wallapop = """
SELECT AVG(precio) AS media_precio FROM datos_wallpop_limpios;
"""

cursor.execute(media_productos_wallapop)
media_precio = cursor.fetchall()[0]  

print("Media de precio:", media_precio)

consulta_insertar = """
INSERT INTO wallapop_interesante (wallpop_id, user_id, titulo, descripcion, categoria_id, precio, 
                                  imagenes, reservado, localizacion, golpe, reacondicionado, 
                                  url_producto, posibilidad_envio)
SELECT wallpop_id, user_id, titulo, descripcion, categoria_id, precio, 
       imagenes, reservado, localizacion, golpe, reacondicionado, 
       url_producto, posibilidad_envio
FROM datos_wallpop_limpios
WHERE precio < %s;
"""

cursor.execute(consulta_insertar, (media_precio,))

conn.commit()

cursor.close()
conn.close()

print("Productos copiados a wallapop_interesante con éxito.")


df_datos_wallapop = swallapop.obtener_datos_wallapop_sql()

archivo_pdf = "datos_wallapop_interesantes.pdf"
swallapop.generar_pdf(df_datos_wallapop, archivo_pdf)

swallapop.enviar_correo_con_adjunto(
    destinatario="almazanlucas@gmail.com",
    asunto="Productos de Wallapop - PS5",
    cuerpo="Adjunto encontrarás el informe de productos de PS5 en Wallapop.",
    archivo_adj=archivo_pdf
)

## Vinted

> A dia de hoy no hay Play Station 5 a la venta en Vinted, no obstante es interesante no perder de vista esta plataforma ya que puede haber en algun momento.

media_vinted = """
SELECT AVG(total_item_price) AS media_precio FROM datos_vinted_filtrados;
"""
cursor.execute(media_vinted)
media_precio_vinted = cursor.fetchone()[0]
print("Media de precio para Vinted:", media_precio_vinted)

insertar_vinted = """
INSERT INTO vinted_interesante (id, title, is_visible, currency, brand_title, url, 
                                promoted, favourite_count, is_favourite, total_item_price, status)
SELECT id, title, is_visible, currency, brand_title, url, 
       promoted, favourite_count, is_favourite, total_item_price, status
FROM datos_vinted_filtrados
WHERE total_item_price < %s;
"""
cursor.execute(insertar_vinted, (media_precio_vinted,))

df_datos_vinted = svinted.obtener_datos_vinted_sql()

archivo_pdf = "datos_vinted_interesantes.pdf"
svinted.generar_pdf(df_datos_vinted, archivo_pdf)

svinted.enviar_correo_con_adjunto(
    destinatario="almazanlucas@gmail.com",
    asunto="Productos de Vinted - PS5",
    cuerpo="Adjunto encontrarás el informe de productos de PS5 en Vinted.",
    archivo_adj=archivo_pdf
)

## Cash Converters

media_cash_converters = """
SELECT AVG(precio) AS media_precio FROM datos_cash_converters;
"""
cursor.execute(media_cash_converters)
media_precio_cash_converters = cursor.fetchone()[0]
print("Media de precio para Cash Converters:", media_precio_cash_converters)

insertar_cash_converters = """
INSERT INTO cash_converters_interesante (titulo, descripcion, precio)
SELECT titulo, descripcion, precio
FROM datos_cash_converters
WHERE precio < %s;
"""
cursor.execute(insertar_cash_converters, (media_precio_cash_converters,))

conn.commit()

cursor.close()
conn.close()

df_datos_cash = scash.obtener_datos_cash_converters_sql()

archivo_pdf = "datos_cash_interesantes.pdf"
scash.generar_pdf(df_datos_cash, archivo_pdf)

scash.enviar_correo_con_adjunto(
    destinatario="almazanlucas@gmail.com",
    asunto="Productos de cash - PS5",
    cuerpo="Adjunto encontrarás el informe de productos de PS5 en cash.",
    archivo_adj=archivo_pdf
)

## Mercado Libre

media_mercado_libre = """
SELECT AVG(precio_eur) AS media_precio FROM datos_mercado_libre;
"""
cursor.execute(media_mercado_libre)
media_precio_mercado_libre = cursor.fetchone()[0]
print("Media de precio para Mercado Libre:", media_precio_mercado_libre)

insertar_mercado_libre = """
INSERT INTO mercado_libre_interesante (titulo, link, precio_eur)
SELECT titulo, link, precio_eur
FROM datos_mercado_libre
WHERE precio_eur < %s;
"""
cursor.execute(insertar_mercado_libre, (media_precio_mercado_libre,))

df_datos_cash = smerc.obtener_datos_mercado_libre_sql()

archivo_pdf = "datos_cash_interesantes.pdf"
smerc.generar_pdf(df_datos_cash, archivo_pdf)

smerc.enviar_correo_con_adjunto(
    destinatario="almazanlucas@gmail.com",
    asunto="Productos de mercado libre - PS5",
    cuerpo="Adjunto encontrarás el informe de productos de PS5 en Mercado Liber.",
    archivo_adj=archivo_pdf
)
